[libcalamares] Deal with Qt5/6 differences in count()

+++ src/libcalamares/modulesystem/RequirementsModel.cpp
@@ -66,7 +66,7 @@ RequirementsModel::reCheckList()
 int
 RequirementsModel::rowCount( const QModelIndex& ) const
 {
-    return m_requirements.count();
+    return static_cast< int >( m_requirements.count() );  // TODO 3.4 use qsizetype
 }
 
 QVariant
